Fix crashes in the ring and jaccl distributed backends
Three unrelated stability issues in the CPU-side distributed backends,
all of which show up as SIGSEGVs on long-running multi-node jobs.

Use-after-free in the collectives. `Encoder::dispatch` only enqueues the
lambda, and `set_input_array`/`set_output_array` are no-ops on the CPU
backend, so nothing keeps the arrays' buffers alive until the task runs.
`cpu::eval` queues a keep-alive holding the inputs, but excludes the
output, which the collectives allocate themselves. Under memory pressure
the allocator can reclaim and reuse the output block before the
collective runs, and the stream thread then writes through a stale
pointer. Capture `data_shared_ptr()` alongside the raw pointers in the
jaccl and ring collectives so the buffers are pinned for the task's
lifetime.

Unvalidated work completions in the jaccl ring. A failed or spurious
completion carries an undefined `wr_id`, and the wire/buff decoded from
it was used unchecked to index the per-wire counters, offsets, limits and
the buffer pools. Out-of-range values produced a wild pointer in
`reduce_op`/`std::copy`. Check `wc.status` and range-check the decoded
indices, then drop the completion and keep draining. The in-flight
counter is decremented before the check so the loop still terminates.

32-bit offset overflow. `read_offset`/`write_offset` in the mesh
all_gather and `n_steps` in the ring all_gather were `int` while indexing
payloads sized by `int64_t`, wrapping past 2GB and ~4GB per rank
respectively. Widen them to `int64_t`.

+++ mlx/distributed/jaccl/lib/jaccl/mesh_impl.h
@@ -214,9 +214,11 @@ class MeshImpl {
 
     // Counters to maintain the state of transfers
     int in_flight = 0;
-    int read_offset = 0;
+    // Byte offsets into a payload sized by the int64_t `total`, so they have
+    // to be 64-bit as well or they wrap past 2GB per rank.
+    int64_t read_offset = 0;
     int completed_send_count[PIPELINE] = {0};
-    int write_offset[MESH_MAX_PEERS] = {0};
+    int64_t write_offset[MESH_MAX_PEERS] = {0};
 
     // Prefill the pipeline
     int buff = 0;
